requirements.txt
# Core dependencies
streamlit>=1.28.0
langchain>=0.1.0
langchainhub>=0.1.0
langchain-chroma>=0.1.0
langchain-community>=0.1.0
langchain-huggingface>=0.1.0
langchain-core>=0.1.0
langchain-experimental>=0.1.0
python-dotenv>=1.0.0
pypdf>=3.0.0
openai>=1.0.0
sentence-transformers>=2.2.0

# Enhanced features
plotly>=5.15.0
pandas>=2.0.0
numpy>=1.24.0
chromadb>=0.4.0

# Additional utilities
typing-extensions>=4.5.0
pydantic>=2.0.0
httpx>=0.25.0

# Development dependencies (optional)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0